pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "nc_cutout"
version = "0.1.0"
description = "JWST NIRCam postage-stamp cutout creator for JADES GOODS-S and GOODS-N data"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Aayush Saxena" }
]
requires-python = ">=3.9"
dependencies = [
    "astropy>=5.0",
    "pandas>=1.4",
    "pyyaml>=6.0",
]

[project.scripts]
nc_cutout = "nc_cutout.cutout:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["nc_cutout*"]
